[Android/Feat] 루틴 삭제 기능 구현

+++ Android/app/src/main/java/com/hogumiwarts/lumos/mapper/RoutineDetailMapper.kt
@@ -6,7 +6,7 @@ import com.hogumiwarts.domain.model.routine.RoutineDetailData
 import com.hogumiwarts.lumos.ui.screens.routine.components.DeviceListType
 import com.hogumiwarts.lumos.ui.screens.routine.components.RoutineItem
 
-fun RoutineDetailData.toRoutineItem(routineId: Int): RoutineItem {
+fun RoutineDetailData.toRoutineItem(routineId: Long): RoutineItem {
     return RoutineItem(
         routineId = routineId,
         routineName = routineName,
